refactor: lib core logic 분리
