+++ src/App.ts
@@ -1,7 +1,7 @@
 import { AppHeader } from "./components/AppHeader";
 import { CurrentNews } from "./components/CurrentNews";
-import { cc } from "./libs/components";
-import { ce, Div } from "./libs/elements";
+import { cc } from "./libs/nodes/components";
+import { ce, Div } from "./libs/nodes/elements";
 import styles from "./App.module.css";
 import { MediaContent } from "./components/MediaContent";
 export const App = () => {
